backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.20.0
apscheduler==3.10.4
sqlalchemy>=2.0.36
qrcode[pil]==7.4.2
Pillow>=10.4.0
python-dotenv==1.0.1
